test: #27 daoのテストケース実装
